+++ ricaun.DI.Tests/DI/Dummies/AbstractDummyService.cs
@@ -1,0 +1,6 @@
+﻿namespace ricaun.DI.Tests.DI.Dummies
+{
+    public abstract class AbstractDummyService
+    {
+    }
+}