+++ ricaun.DI.Tests/DI/Dummies/CircularService1.cs
@@ -1,0 +1,10 @@
+﻿namespace ricaun.DI.Tests.DI.Dummies
+{
+    public class CircularService1
+    {
+        public CircularService1(CircularService1 circularService1)
+        {
+
+        }
+    }
+}