+++ ricaun.DI.Tests/DI/Dummies/CircularService2.cs
@@ -1,0 +1,9 @@
+﻿namespace ricaun.DI.Tests.DI.Dummies
+{
+    public class CircularService2
+    {
+        public CircularService2(DummyService dummyService, CircularService3 circularService3)
+        {
+        }
+    }
+}