+++ ricaun.DI.Tests/DI/Dummies/CircularService3.cs
@@ -1,0 +1,9 @@
+﻿namespace ricaun.DI.Tests.DI.Dummies
+{
+    public class CircularService3
+    {
+        public CircularService3(CircularService2 circularService2)
+        {
+        }
+    }
+}