+++ ricaun.DI.Tests/DI/Dummies/DependentService.cs
@@ -1,0 +1,9 @@
+﻿namespace ricaun.DI.Tests.DI.Dummies
+{
+    public class DependentService
+    {
+        public DependentService(DummyService dummyService)
+        {
+        }
+    }
+}