+++ ricaun.DI.Tests/DI/Dummies/DummyHiddenConstructor.cs
@@ -1,0 +1,7 @@
+﻿namespace ricaun.DI.Tests.DI.Dummies
+{
+    public class DummyHiddenConstructor
+    {
+        private DummyHiddenConstructor() { }
+    }
+}