+++ ricaun.DI.Tests/DI/Dummies/DummyService1.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ricaun.DI.Tests.DI.Dummies
+{
+    public class DummyService1
+    {
+        public DummyService1()
+        {
+        }
+    }
+}