+++ ricaun.DI.Tests/DI/Dummies/IDummyService.cs
@@ -1,0 +1,6 @@
+﻿namespace ricaun.DI.Tests.DI.Dummies
+{
+    public interface IDummyService
+    {
+    }
+}