+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test `Transient`
 - Test `DI` and `Dispose`
 - Test `Abstract` / `Interface` throws
+- Test `ContainerShould` with `Dummies`
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0